Add NewAppendableFile for win32 environment
Implement the new NewAppendableFile call for the win32 environment.
This is used to open an existing file for writing at the end,
whereas NewWritableFile is supposed to delete/truncate if the file
already exists. If the file does not exist the two are equivalent.

Note that all writing in LevelDB is appending, so the only difference
between a WritableFile and AppendableFile is in construction and opening.
Thus, NewAppendableFile simply returns a WritableFile interface.